Redis IntegrationTest 작성

+++ src/main/java/com/sjy/dayontest/repository/StudentScoreRepository.java
@@ -3,5 +3,5 @@ package com.sjy.dayontest.repository;
 import com.sjy.dayontest.model.StudentScore;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentScoreRepository extends JpaRepository<StudentScore, Integer> {
+public interface StudentScoreRepository extends JpaRepository<StudentScore, Long> {
 }
